--- Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Toast/Controls Visuals/Basic UI/RichTextBox/BasicNotification.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Krypton.Toolkit.Suite.Extended.Toast
-{
-    public partial class BasicNotification : KryptonForm
-    {
-        public BasicNotification()
-        {
-            InitializeComponent();
-        }
-    }
-}
