+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Toast/UI/Basic UI/RichTextBox/BasicNotificationWithProgressBar.resx
