+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Toast/UI/User Response/RichTextBox/BasicNotificationWithUserResponseAndProgressBar.resx
