+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Toast/UI/User Response/Wrapped Label/BasicNotificationWithUserResponseWrappedLabelAndProgressBar.cs
